requirements.txt
langchain==0.3.26
langchain-community==0.3.26
langchain-openai==0.3.28
faiss-cpu==1.13.2
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
fastapi==0.115.0
uvicorn==0.30.6
pytest==8.3.3
